Splitted map.test.js (internal-1223)
